Prepare for upcoming breaking changes in FastBoot 1.0
