--- addon/instance-initializers/browser/responsive-meta.js
@@ -1,14 +1,0 @@
-import ResponsiveImage from '../../services/responsive-image';
-import Ember from 'ember';
-
-export function initialize(/* appInstance */) {
-  let meta = JSON.parse(Ember.$('#ember_responsive_image_meta').text());
-  ResponsiveImage.reopen({
-    meta
-  });
-}
-
-export default {
-  name: 'responsive-meta',
-  initialize
-};
