--- app/instance-initializers/browser/responsive-meta.js
@@ -1,1 +1,0 @@
-export { default, initialize } from 'ember-responsive-image/instance-initializers/browser/responsive-meta';
