+++ app/instance-initializers/responsive-meta.js
@@ -1,0 +1,1 @@
+export { default, initialize } from 'ember-responsive-image/instance-initializers/responsive-meta';
