+++ package.json
@@ -25,7 +25,6 @@
     "broccoli-stew": "^1.4.0",
     "chalk": "^1.1.1",
     "ember-cli-babel": "^5.1.7",
-    "fastboot-filter-initializers": "^0.0.2",
     "fs-extra": "^0.26.5",
     "gm": "^1.21.1",
     "mkdirp": "^0.5.1",
