+++ tests/integration/components/responsive-image-test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { initialize } from 'ember-responsive-image/instance-initializers/browser/responsive-meta';
+import { initialize } from 'ember-responsive-image/instance-initializers/responsive-meta';
 import {
   setupComponentTest,
   it
